controller changed to flask views
