+++ ckanext/issues/controller/__init__.py
@@ -1,3 +1,3 @@
-from .controller import IssueController
-from .moderation import ModerationController
-from .moderation import CommentModerationController
+from .issues import *
+from .moderation import *
+
