+++ ckanext/issues/plugin.py
@@ -7,6 +7,9 @@ log = getLogger(__name__)
 import ckan.plugins as p
 from ckan.plugins import implements, toolkit
 
+from ckanext.issues.controller.issues import issues
+from ckanext.issues.controller.moderation import moderation
+
 # Imports are done in methods to speed up paster.
 # Please don't move back up to here.
 
@@ -16,19 +19,40 @@ class IssuesPlugin(p.SingletonPlugin):
     """
     implements(p.IConfigurer, inherit=True)
     implements(p.ITemplateHelpers, inherit=True)
-    implements(p.IRoutes, inherit=True)
+    # implements(p.IRoutes, inherit=True)
     implements(p.IActions)
     implements(p.IAuthFunctions)
 
     # IConfigurer
-
     def update_config(self, config):
         toolkit.add_template_directory(config, 'templates')
         toolkit.add_public_directory(config, 'public/css')
         toolkit.add_resource('public/scripts', 'ckanext_issues')
 
-    # ITemplateHelpers
 
+    implements(p.IClick)
+    # IClick
+    def get_commands(self):
+        """CLI commands - Creates or Updated issues data tables"""       
+        import click
+
+        @click.command()
+        def init_db():
+            """Creates issues data tables"""
+            from ckanext.issues.model import setup
+            setup()
+
+        @click.command()
+        def upgrade_db():
+            """Updates issues data tables"""
+            from ckanext.issues.model import upgrade
+            upgrade()
+            print('Issues tables are up to date')
+
+        return [init_db, upgrade_db]
+
+    
+    # ITemplateHelpers
     def get_helpers(self):
         from ckanext.issues.lib import util, helpers
         return {
@@ -50,77 +74,13 @@ class IssuesPlugin(p.SingletonPlugin):
                 helpers.issues_users_who_reported_issue,
         }
 
-    # IRoutes
 
-    def before_map(self, map):
-        from ckan.config.routing import SubMapper
-
-        controller_name = 'ckanext.issues.controller:IssueController'
-        with SubMapper(map, controller=controller_name) as m:
-            m.connect('issues_dataset',
-                      '/dataset/:dataset_id/issues',
-                      action='dataset',
-                      ckan_icon='warning-sign')
-            m.connect('issues_new',
-                      '/dataset/:dataset_id/issues/new',
-                      action='new')
-            m.connect('issues_edit',
-                      '/dataset/:dataset_id/issues/:issue_number/edit',
-                      action='edit')
-            m.connect('issues_delete',
-                      '/dataset/:dataset_id/issues/:issue_number/delete',
-                      action='delete')
-            m.connect('issues_assign',
-                      '/dataset/:dataset_id/issues/:issue_number/assign',
-                      action='assign')
-            m.connect('issues_comments',
-                      '/dataset/:dataset_id/issues/:issue_number/comments',
-                      action='comments')
-            m.connect('issues_report',
-                      '/dataset/:dataset_id/issues/:issue_number/report',
-                      action='report'),
-            m.connect('issues_report_clear',
-                      '/dataset/:dataset_id/issues/:issue_number/report_clear',
-                      action='report_clear'),
-            m.connect('issues_comment_report',
-                      '/dataset/:dataset_id/issues/:issue_number/comment/:comment_id/report',
-                      action='report_comment'),
-            m.connect('issues_comment_report_clear',
-                      '/dataset/:dataset_id/issues/:issue_number/comment/:comment_id/report_clear',
-                      action='comment_report_clear'),
-            m.connect('add_issue_with_resource',
-                      '/dataset/:dataset_id/issues/new/:resource_id',
-                      action='add')
-            m.connect('issues_show',
-                      '/dataset/:dataset_id/issues/:issue_number',
-                      action='show')
-            m.connect('all_issues_page', '/issues', action='all_issues_page')
-            m.connect('issues_for_organization',
-                      '/organization/:org_id/issues',
-                      action='issues_for_organization')
-
-        moderation = 'ckanext.issues.controller:ModerationController'
-        with SubMapper(map, controller=moderation) as m:
-            m.connect('issues_moderate_reported_issues',
-                      '/organization/:organization_id/issues/reported',
-                      action='all_reported_issues')
-            m.connect('issues_moderate',
-                      '/organization/:organization_id/issues/moderate',
-                      action='moderate')
-
-        moderation = 'ckanext.issues.controller:CommentModerationController'
-        with SubMapper(map, controller=moderation) as m:
-            m.connect('issues_moderate_reported_comments',
-                      '/organization/:organization_id/issues/reported_comments',
-                      action='reported_comments')
-            m.connect('issues_moderate_comment',
-                      '/organization/:organization_id/issues/moderate_comment',
-                      action='moderate')
-
-        return map
+    p.implements(p.IBlueprint)
+    # IBlueprint
+    def get_blueprint(self):
+        return [issues, moderation]
 
     # IActions
-
     def get_actions(self):
         import ckanext.issues.logic.action as action
 
@@ -129,7 +89,6 @@ class IssuesPlugin(p.SingletonPlugin):
                     if callable(function))
 
     # IAuthFunctions
-
     def get_auth_functions(self):
         import ckanext.issues.auth as auth
         return {
@@ -144,3 +103,72 @@ class IssuesPlugin(p.SingletonPlugin):
             'issue_report_clear': auth.issue_report_clear,
             'issue_comment_search': auth.issue_comment_search,
         }
+
+    # IRoutes
+
+    # def before_map(self, map):
+    #     from ckan.config.routing import SubMapper
+
+    #     controller_name = 'ckanext.issues.controller:IssueController'
+    #     with SubMapper(map, controller=controller_name) as m:
+    #         m.connect('issues_dataset',
+    #                   '/dataset/:dataset_id/issues',
+    #                   action='dataset',
+    #                   ckan_icon='warning-sign')
+    #         m.connect('issues_new',
+    #                   '/dataset/:dataset_id/issues/new',
+    #                   action='new')
+    #         m.connect('issues_edit',
+    #                   '/dataset/:dataset_id/issues/:issue_number/edit',
+    #                   action='edit')
+    #         m.connect('issues_delete',
+    #                   '/dataset/:dataset_id/issues/:issue_number/delete',
+    #                   action='delete')
+    #         m.connect('issues_assign',
+    #                   '/dataset/:dataset_id/issues/:issue_number/assign',
+    #                   action='assign')
+    #         m.connect('issues_comments',
+    #                   '/dataset/:dataset_id/issues/:issue_number/comments',
+    #                   action='comments')
+    #         m.connect('issues_report',
+    #                   '/dataset/:dataset_id/issues/:issue_number/report',
+    #                   action='report'),
+    #         m.connect('issues_report_clear',
+    #                   '/dataset/:dataset_id/issues/:issue_number/report_clear',
+    #                   action='report_clear'),
+    #         m.connect('issues_comment_report',
+    #                   '/dataset/:dataset_id/issues/:issue_number/comment/:comment_id/report',
+    #                   action='report_comment'),
+    #         m.connect('issues_comment_report_clear',
+    #                   '/dataset/:dataset_id/issues/:issue_number/comment/:comment_id/report_clear',
+    #                   action='comment_report_clear'),
+    #         m.connect('add_issue_with_resource',
+    #                   '/dataset/:dataset_id/issues/new/:resource_id',
+    #                   action='add')
+    #         m.connect('issues_show',
+    #                   '/dataset/:dataset_id/issues/:issue_number',
+    #                   action='show')
+    #         m.connect('all_issues_page', '/issues', action='all_issues_page')
+    #         m.connect('issues_for_organization',
+    #                   '/organization/:org_id/issues',
+    #                   action='issues_for_organization')
+
+    #     moderation = 'ckanext.issues.controller:ModerationController'
+    #     with SubMapper(map, controller=moderation) as m:
+    #         m.connect('issues_moderate_reported_issues',
+    #                   '/organization/:organization_id/issues/reported',
+    #                   action='all_reported_issues')
+    #         m.connect('issues_moderate',
+    #                   '/organization/:organization_id/issues/moderate',
+    #                   action='moderate')
+
+    #     moderation = 'ckanext.issues.controller:CommentModerationController'
+    #     with SubMapper(map, controller=moderation) as m:
+    #         m.connect('issues_moderate_reported_comments',
+    #                   '/organization/:organization_id/issues/reported_comments',
+    #                   action='reported_comments')
+    #         m.connect('issues_moderate_comment',
+    #                   '/organization/:organization_id/issues/moderate_comment',
+    #                   action='moderate')
+
+    #     return map
